+++ HowsFront/src/pages/Home/Products/Main/Detail/DeliveryInfo/DeliveryInfo.module.css
@@ -1,0 +1,73 @@
+.container {
+    padding: 50px 20px;
+}
+.container h4, .container h3 {
+    margin: 10px 0;
+}
+.container > div {
+    margin-bottom: 100px;
+}
+.line{
+    border-bottom: 1px solid #bcbcbc25; 
+    padding: 5px;
+}
+.container > div > div > div {
+    display: flex;
+}
+.container > div > div > div > div:first-child{
+    flex: 2;
+}
+.container > div > div > div > div:last-child{
+    flex: 8;
+}
+.container > div > p {
+    color: var(--hows-point-color-darker);
+    font-size: 14px; 
+    font-weight: bold;
+    margin-bottom: 10px;
+}
+.container span {
+    font-size: 12px;
+    color: var(--hows-point-color-darker);
+    font-weight: bold;
+}
+.container >div:nth-child(2) > h4 {
+    margin-top: 50px;
+}
+.container > div > ol li{
+    margin-bottom: 10px;
+}
+
+/* 도비즈 상세정보 */
+.card {
+    padding: 20px;
+    background-color: #f9f9f9;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    max-width: 800px; /* 최대 너비를 설정하여 너무 넓어지지 않도록 */
+    margin: 0 auto; /* 중앙 정렬 */
+}
+
+.card_line {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 10px;
+    padding: 10px 0;
+    /* border-bottom: 1px solid #ddd; */
+}
+
+.label {
+    font-weight: bold;
+    color: #333;
+    width: 30%; /* 제목과 값의 너비를 비율로 설정 */
+}
+
+.value {
+    color: #555;
+    width: 70%; /* 제목과 값의 너비를 비율로 설정 */
+}
+
+/* 마지막 항목의 경계선 제거 */
+.line:last-child {
+    border-bottom: none;
+}
