+++ HowsServer/src/main/java/com/hows/product/controller/LikesController.java
@@ -21,7 +21,7 @@ public class LikesController {
 	@Autowired
 	private LikesService likesServ;
 	
-	// 좋아요 추가
+	// 상품 좋아요 추가
     @PostMapping("/insert")
     public ResponseEntity<String> addLike(@RequestBody Map<String, Object> requestData) 
     throws Exception{
@@ -32,7 +32,7 @@ public class LikesController {
          return ResponseEntity.ok("Like");
     }
     
-    // 좋아요 취소
+    // 상품 좋아요 취소
     @DeleteMapping("/delete")
     public ResponseEntity<String> removeLike(@RequestBody Map<String, Object> requestData) 
     throws Exception{
@@ -43,7 +43,7 @@ public class LikesController {
         return ResponseEntity.ok("Like removed");
     }
     
-    // 좋아요 개수 조회
+    // 상품 좋아요 개수 조회
     @GetMapping("/count")
     public ResponseEntity<Integer> getLikeCount(
     		@RequestParam("product_seq") int product_seq) 
@@ -52,7 +52,7 @@ public class LikesController {
         return ResponseEntity.ok(likeCount);
     }
     
-    // 좋아요 확인
+    // 상품 좋아요 확인
     @GetMapping("/check")
     public ResponseEntity<Boolean> checkLikeStatus(
     		@RequestParam("product_seq") int product_seq,
@@ -67,6 +67,48 @@ public class LikesController {
 
     	
         boolean isLiked = likesServ.isLiked(product_seq, member_id);
+        return ResponseEntity.ok(isLiked);
+    }
+    
+    
+    // 리뷰 좋아요 추가
+    @PostMapping("/review/insert")
+    public ResponseEntity<String> addReviewLike(@RequestBody Map<String, Object> requestData) 
+    throws Exception {
+        int review_seq = Integer.parseInt(requestData.get("review_seq").toString());
+        String member_id = (String) requestData.get("member_id");
+
+        likesServ.addReviewLike(review_seq, member_id);
+        return ResponseEntity.ok("Review liked");
+    }
+
+    // 리뷰 좋아요 취소
+    @DeleteMapping("/review/delete")
+    public ResponseEntity<String> removeReviewLike(@RequestBody Map<String, Object> requestData) 
+    throws Exception {
+        int review_seq = Integer.parseInt(requestData.get("review_seq").toString());
+        String member_id = (String) requestData.get("member_id");
+
+        likesServ.removeReviewLike(review_seq, member_id);
+        return ResponseEntity.ok("Review like removed");
+    }
+
+    // 리뷰 좋아요 개수 조회
+    @GetMapping("/review/count")
+    public ResponseEntity<Integer> getReviewLikeCount(
+        @RequestParam("review_seq") int review_seq) 
+    throws Exception {
+        int likeCount = likesServ.getReviewLikeCount(review_seq);
+        return ResponseEntity.ok(likeCount);
+    }
+
+    // 리뷰 좋아요 확인
+    @GetMapping("/review/check")
+    public ResponseEntity<Boolean> checkReviewLikeStatus(
+        @RequestParam("review_seq") int review_seq,
+        @RequestParam("member_id") String member_id) 
+    throws Exception {
+        boolean isLiked = likesServ.isReviewLiked(review_seq, member_id);
         return ResponseEntity.ok(isLiked);
     }
